pyproject.toml
[project]
name = "ai-blog-platform"
version = "0.1.0"
description = "AI-powered blog platform that scrapes news, papers, and articles to generate high-quality blog posts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "supabase>=2.3.0",
    "httpx>=0.26.0",
    "google-generativeai>=0.3.0",
    "apscheduler>=3.10.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-slugify>=8.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
